+++ packages/client/src/pages/Home/Home.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react';
 
 import { Header } from '@components/Header';
 
-import { About } from './parts/About';
-import { Hero } from './parts/Hero';
+import { About } from './components/About';
+import { Hero } from './components/Hero';
 
 export const Home: FC = () => {
   return (
