+++ packages/client/src/pages/Home/components/About/About.tsx
