+++ packages/client/src/pages/Home/components/About/index.ts
