+++ packages/client/src/pages/Home/components/Hero/Hero.tsx
