+++ packages/client/src/pages/Home/components/Hero/index.ts
