--- packages/client/src/pages/Leaderboard/LeaderBoard.css
@@ -1,9 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer components {
-  .leader-board-item {
-    @apply py-4 px-5 w-full flex flex-col sm:grid sm:grid-cols-[min-content_40px_30%_1fr_1fr] items-center gap-x-5;
-  }
-}
